requirements.txt
streamlit
pandas
Pillow
qrcode
pymupdf
xlsxwriter
openpyxl